Narrowed scope of file by breaking into smaller chunks
