+++ src/route/consts.py
@@ -1,0 +1,3 @@
+"""Routing module constants"""
+
+__default_timeout__: int = 30
